Keep monitoring response close to functions
